Game/requirement.txt
asgiref==3.7.2
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
backports.zoneinfo==0.2.1
blinker==1.4
bottle==0.12.25
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
cloud-init==23.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
Cython==0.29.14
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.7
djangorestframework==3.14.0
entrypoints==0.3
GeoIP==1.3.2
hkdf==0.0.3
httplib2==0.14.0
humanize==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libvirt-python==6.1.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
magic-wormhole==0.11.2
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
norminette==3.3.54
oauthlib==3.1.0
pexpect==4.6.0
protobuf==3.6.1
psutil==5.5.1
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pypng==0.0.20
PyQRCode==1.2.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.3
spake2==0.8
speedtest-cli==2.1.2
sqlparse==0.4.4
ssh-import-id==5.10
supervisor==4.1.0
systemd-python==234
tqdm==4.30.0
Twisted==18.9.0
txaio==2.10.0
txtorcon==19.0.0
typing-extensions==4.8.0
u-msgpack-python==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wsaccel==0.6.2
xkit==0.0.0
zipp==1.0.0
zope.interface==4.7.1
